fixed codeforce timezone and seperated tables
