+++ public/js/contest_countdown.js
@@ -1,13 +1,12 @@
-var table = document.getElementById("display-all-contests");
-var rows = table.getElementsByTagName("tr");
+var rows = document.querySelectorAll(".display-all-contests tr");
 var deadlineArr = [];
-for (var i = 0; i < rows.length - 1; i++) {
+for (var i = 0; i < rows.length; i++) {
     var tr = document.getElementById(i.toString(10));
     var dt = tr.getElementsByTagName("td");
     deadlineArr[i] = new Date(dt[2].innerText);
 }
 var x = setInterval(function () {
-    for (var i = 0; i < rows.length - 1; i++) {
+    for (var i = 0; i < rows.length; i++) {
         var deadline = new Date(deadlineArr[i]).getTime();
         var now = new Date().getTime();
         var t = deadline - now;
